+++ plutus-numeric/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 4.0 -- 2021-12-07
+
+### Added
+
+* New and improved `Rational` type, which is more efficient than the one
+  provided by Plutus.
+
 ## 3.0 -- 2021-12-04
 
 ### Removed